+++ packages/react/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @vleap/warps-react
 
+## 0.0.5-alpha.7
+
+### Patch Changes
+
+- Updated dependencies [b978b7a]
+  - @vleap/warps@2.0.0-alpha.7
+
+## 0.0.5-alpha.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @vleap/warps@2.0.0-alpha.6
+
+## 0.0.5-alpha.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @vleap/warps@2.0.0-alpha.5
+
+## 0.0.5-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [f806686]
+  - @vleap/warps@2.0.0-alpha.4
+
+## 0.0.5-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [60cd189]
+  - @vleap/warps@2.0.0-alpha.3
+
+## 0.0.5-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [f55412d]
+  - @vleap/warps@2.0.0-alpha.2
+
+## 0.0.5-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [7d50e91]
+  - @vleap/warps@2.0.0-alpha.1
+
+## 0.0.5-alpha.0
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @vleap/warps@2.0.0-alpha.0
+
 ## 0.0.4
 
 ### Patch Changes